+++ src/linter.rs
@@ -251,6 +251,8 @@ impl Linter {
                 | "random_float"
                 | "println"
                 | "exec"
+                | "_change_dir"
+                | "_run_process"
                 | "file_exists"
                 | "is_file"
                 | "is_dir"
@@ -474,6 +476,8 @@ impl Linter {
         self.functions.insert("exit".to_string());
         self.functions.insert("args".to_string());
         self.functions.insert("cwd".to_string());
+        self.functions.insert("_change_dir".to_string());
+        self.functions.insert("_run_process".to_string());
         self.functions.insert("now".to_string());
         self.functions.insert("timestamp".to_string());
         self.functions.insert("sleep".to_string());